+++ src/main/java/org/sosly/workersplus/entities/workers/Porter.java
@@ -31,8 +31,7 @@ import org.sosly.workersplus.networking.PacketHandler;
 import org.sosly.workersplus.networking.clientbound.UpdateHireScreen;
 import org.sosly.workersplus.networking.serverbound.OpenHireGUI;
 import org.sosly.workersplus.networking.serverbound.OpenWorkerGUI;
-import org.sosly.workersplus.tasks.CollectionTask;
-import org.sosly.workersplus.tasks.DeliveryTask;
+import org.sosly.workersplus.tasks.PorterTask;
 import org.sosly.workersplus.tasks.TaskCoordinator;
 
 import java.util.List;
@@ -41,8 +40,7 @@ import java.util.function.Predicate;
 
 public class Porter extends AbstractWorkerEntity {
     private WorkerRelationships relationships;
-    private DeliveryTask delivery;
-    private CollectionTask collection;
+    private PorterTask porterTask;
     private TaskCoordinator taskCoordinator;
 
     public Porter(EntityType<? extends AbstractWorkerEntity> entityType, Level world) {
@@ -58,11 +56,8 @@ public class Porter extends AbstractWorkerEntity {
         CompoundTag relationships = this.relationships.save();
         nbt.put("relationships", relationships);
 
-        CompoundTag deliveryData = this.delivery.save();
-        nbt.put("delivery", deliveryData);
-
-        CompoundTag collectionData = this.collection.save();
-        nbt.put("collection", collectionData);
+        CompoundTag porterData = this.porterTask.save();
+        nbt.put("porter", porterData);
     }
 
     @Override
@@ -74,14 +69,9 @@ public class Porter extends AbstractWorkerEntity {
             this.relationships.load(relationships);
         }
 
-        if (nbt.contains("delivery")) {
-            CompoundTag deliveryData = nbt.getCompound("delivery");
-            this.delivery.load(deliveryData);
-        }
-
-        if (nbt.contains("collection")) {
-            CompoundTag collectionData = nbt.getCompound("collection");
-            this.collection.load(collectionData);
+        if (nbt.contains("porter")) {
+            CompoundTag porterData = nbt.getCompound("porter");
+            this.porterTask.load(porterData);
         }
     }
 
@@ -95,12 +85,8 @@ public class Porter extends AbstractWorkerEntity {
             taskCoordinator = new TaskCoordinator();
         }
         
-        if (delivery == null) {
-            delivery = new DeliveryTask(this, taskCoordinator);
-        }
-
-        if (collection == null) {
-            collection = new CollectionTask(this, taskCoordinator);
+        if (porterTask == null) {
+            porterTask = new PorterTask(this, taskCoordinator);
         }
 
         if (relationships == null) {
@@ -114,24 +100,15 @@ public class Porter extends AbstractWorkerEntity {
 
         registerDependencies();
 
-        // Delivery task goals
-        this.goalSelector.addGoal(10, new StartTaskGoal(this, collection));
-        this.goalSelector.addGoal(5, new PutItemsInContainerGoal(this, delivery));
-        this.goalSelector.addGoal(5, new GetItemsFromContainerGoal(this, delivery));
-        this.goalSelector.addGoal(5, new MoveToDestinationGoal(this, delivery, 2.5D));
-        this.goalSelector.addGoal(5, new AssessWorkerNeedsGoal(this, delivery));
-        this.goalSelector.addGoal(5, new TargetKnownWorkerGoal(this, delivery, relationships));
-        this.goalSelector.addGoal(10, new EndTaskGoal(delivery));
-
-        // Collection task goals
-        this.goalSelector.addGoal(10, new StartTaskGoal(this, delivery));
-        this.goalSelector.addGoal(5, new PutItemsInContainerGoal(this, collection));
-        this.goalSelector.addGoal(5, new GetItemsFromContainerGoal(this, collection));
-        this.goalSelector.addGoal(5, new VerifyCollectionGoal(this, collection));
-        this.goalSelector.addGoal(5, new MoveToDestinationGoal(this, collection, 2.5D));
-        this.goalSelector.addGoal(5, new AssessWorkerExcessGoal(this, collection));
-        this.goalSelector.addGoal(5, new TargetKnownWorkerGoal(this, collection, relationships));
-        this.goalSelector.addGoal(10, new EndTaskGoal(collection));
+        // Porter task goals
+        this.goalSelector.addGoal(10, new StartTaskGoal(this, porterTask));
+        this.goalSelector.addGoal(5, new PutItemsInContainerGoal(this, porterTask));
+        this.goalSelector.addGoal(5, new GetItemsFromContainerGoal(this, porterTask));
+        this.goalSelector.addGoal(5, new MoveToDestinationGoal(this, porterTask, 2.5D));
+        this.goalSelector.addGoal(5, new AssessWorkerNeedsGoal(this, porterTask));
+        this.goalSelector.addGoal(5, new AssessWorkerExcessGoal(this, porterTask));
+        this.goalSelector.addGoal(5, new TargetKnownWorkerGoal(this, porterTask, relationships));
+        this.goalSelector.addGoal(10, new EndTaskGoal(porterTask));
 
         // General worker goals (lowest priority)
         this.goalSelector.addGoal(1, new PanicGoal(this, 1.3D));